Example use upload
This file is meant to be a practical use case of the previous file entry. This is using multiple sequencial fgets functions to help create a character for another project. It uses conditional loops for the player to decide to continue with the game or to restart their character. This file does not have much practical error handling yet, just basic out of scope handling. Will attempt to merge this file with more error and selection options for the game myMUD being worked on. This will then function as a template for the next iteration.